make get store reusable
